Changed fixed recv buf to 4 * 1024.

+++ src/main/java/com/github/cao/awa/kalmia/network/packet/inbound/ping/TryPingResponsePacket.java
@@ -25,6 +25,6 @@ public class TryPingResponsePacket extends PingPacket {
 
     @Override
     public void inbound(UnsolvedRequestRouter router, PingHandler handler) {
-        long responseMillions = TimeUtil.processNano(startTime()) / 1000000;
+        double responseMillions = TimeUtil.processNano(startTime()) / 1000000D;
     }
 }
